error updated

+++ src/utils/cloudinary.js
@@ -12,15 +12,19 @@ const uploadOnCloudinary = async (localFilePath) => {
     if (!localFilePath) return null;
     //upload file on cloudinary
 
-    const response = await cloudinary.v2.uploader(localFilePath, {
+    const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
 
     //file has been uploaded successfully
-    console.log("file is uploaded on cloudinary", response.url);
+    // console.log("file is uploaded on cloudinary", response.url);
+
+    fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath); //remove the locally saved tempory file as the upload operation got failed
+    console.log("Cloudinary upload error", error);
+
+    if (localFilePath) fs.unlinkSync(localFilePath); //remove the locally saved tempory file as the upload operation got failed
 
     return null;
   }
